// vitals/core/src/models/product-grid-inputs.model.ts
export interface ProductGridInputs {
  config?: string;
  headingText: string;
  categoryAlias: string;
  categoryFilter: string;
  productUrlTemplate: string;
  maxItems: string;
  columns: string;
  showFilters: string;
  sortOrder: string;
  sortBy: string;
  layout: string;
  theme: string;
  variant: string;
  variantKey: string;
}